DEV_PATH/limba/requirements.txt
django
psycopg2-binary
gunicorn
gevent
djangorestframework
django-filter
django-cors-headers
djoser
celery==5.0.5
redis
Pillow

